Module Aws

This module contains code that is common across generated AWS libraries as well as pure code that is useful in writing runtime implementations for the AWS libraries.

module Error : sig ... end

This contains errors that are returned from AWS api calls.

module Request : sig ... end

This contains the http-library agnostic representation of requests to be used by runtime implementations.

module type Call = sig ... end

All AWS api operations should have type Call. Runtime implementation should take as input modules of type Call.

type ('i, 'o, 'e) call = (module Call with type error = 'e and type input = 'i and type output = 'o)

module Time : sig ... end

This module provides parsing / formatting for AWS style timestamps. For example: 2013-05-24T21:15:31.000Z It does not parse the milliseconds (it just truncates them).

module Query : sig ... end

This module is used to produce the nested url query structure used in AWS api requests.

module Xml : sig ... end

This module contains helpers used for XML parsing. It wraps Ezxmlm and adds helpers.

module Json : sig ... end

This module contains a Json type (compatible with Yojson.Basic.t) and helpers.

module Util : sig ... end

This module contains various helpers used in generated code.

module Signing : sig ... end

This module contains the V2 and V4 Authorization header AWS signature algorithm.

module BaseTypes : sig ... end

This module contains base case types for the various datatypes used as input or output by AWS api calls.

module Endpoints : sig ... end