Module Aws.BaseTypes

This module contains base case types for the various datatypes used as input or output by AWS api calls.

module type Base = sig ... end

module Unit : Base with type t = unit

module String : Base with type t = string

module Blob : Base with type t = string

module Boolean : Base with type t = bool

module Integer : Base with type t = int

module Long : Base with type t = int

module Double : Base with type t = float

module Float : Base with type t = float

module DateTime : Base with type t = CalendarLib.Calendar.t