Module Aws.Request

This contains the http-library agnostic representation of requests to be used by runtime implementations.

type meth = [

| `DELETE

| `GET

| `HEAD

| `OPTIONS

| `CONNECT

| `TRACE

| `Other of string

| `PATCH

| `POST

| `PUT

]

HTTP methods. This is (intentionally) compatible with Cohttp.Code.meth to make the cohttp-based runtime implementation easier.

val string_of_meth : meth -> string

Produces a string from the method. This is needed for AWS request signing.

type headers = (string * string) list

HTTP headers.

type signature_version =

| V4

| V2

| S3
Signature version

type t = meth * Uri.t * headers

A request is a method, a uri, and a list of headers.