Module Aws.Xml

This module contains helpers used for XML parsing. It wraps Ezxmlm and adds helpers.

exception RequiredFieldMissing of string

This is thrown when parsing XML responses. It will be caught in response handlers (M.of_http for api call M) and turned into an Error.t

val member : string -> Ezxmlm.nodes -> Ezxmlm.nodes option

This function is identical to Ezxmlm.member except that in the case of an Ezxmlm.Tag_not_found exception it returns None.

val members : string -> Ezxmlm.nodes -> Ezxmlm.nodes list

This function is identical to Ezxmlm.members except in the case of an Ezxmlm.Tag_not_found exception it returns .

val required : string -> 'a option -> 'a

This takes an error message and option time, and throws a RequiredFieldMissing exception (not exported) if the value is None.