Module Aws.Signing

This module contains the V2 and V4 Authorization header AWS signature algorithm.

module Hash : sig ... end

val encode_query : (string * string list) list -> string

val sign_request : access_key:string -> secret_key:string -> service:string -> region:string -> Request.t -> Request.t

Given a service, region, and request, produce a new request with an Authorization header constructed according to the V4 Signing algorithm. This code is a direct translation of the reference implementation from python provided at:

http://docs.aws.amazon.com/general/latest/gr/sigv4-signed-request-examples.html

Note: This requires AWS_ACCESS_KEY and AWS_SECRET_KEY environment variables to be set.

Also: Your system time must be accurate. If you are getting invalid authorization errors, check your system time!

val sign_v2_request : access_key:string -> secret_key:string -> service:string -> region:string -> Request.t -> Request.t

Given a service, region, and request, produce a new request with an Authorization header constructed according to the V2 Signing algorithm. This code is a direct translation of the reference description:

https://docs.aws.amazon.com/general/latest/gr/signature-version-2.html

Note: This requires AWS_ACCESS_KEY and AWS_SECRET_KEY environment variables to be set.

Also: Your system time must be accurate. If you are getting invalid authorization errors, check your system time!